Moved global ruby scripts to scripts folder
