+++ doc/updating_godot_version.md
@@ -5,5 +5,5 @@ When updating to a newer version of Godot the version number needs
 to be changed in the following files:
 
 - `doc/setup_instructions.md`
-- `godot_version.rb`
+- `scripts/godot_version.rb`
 - `docker/ci/Dockerfile`
