+++ scripts/godot_version.rb
