+++ scripts/install_godot_templates_linux.rb
@@ -7,7 +7,7 @@ require 'httparty'
 require_relative 'bootstrap_rubysetupsystem'
 require_relative 'RubySetupSystem/RubyCommon'
 
-require_relative 'godot_version'
+require_relative 'scripts/godot_version'
 
 INSTALL_TARGET = File.join(File.expand_path('~'), '.local/share/godot/templates/',
                            GODOT_VERSION_FULL).freeze
